mergeui-1

+++ src/main/java/com/fp/repository/FraudIntakeObjectRepository.java
@@ -4,7 +4,9 @@ import com.fp.model.FraudIntakeObject;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface FraudIntakeObjectRepository extends JpaRepository<FraudIntakeObject, Long> {
+public interface FraudIntakeObjectRepository extends JpaRepository<FraudIntakeObject, BigDecimal> {
 
 }
